コモンウェルスエンジニアーズレポート/自動化システム/requirements.txt
requests>=2.31.0
PyJWT[crypto]>=2.8.0
python-pptx>=0.6.21
python-dateutil>=2.8.0


